This is an embrional stage and this commit does not constitute a working application. All features not yet implemented are disabled in the interface. As of right now the only thing to see in this application is the initial window and the layout of the wallet frame inside it.
Contact list is almost completely done.
Reading, writing and displaying of contacts works, the only things missing are new and edit contact.

Since the use can't add or edit contacts right now an example json file with a bunch of entries is
present in the main folder and should be copied to "< userdir >/.Algorand Wallet Manager/" and renamed to
contacts.json.